requirements.txt
tensorflow
flask
numpy
pillow
scikit-learn
matplotlib